# docs/openapi.yaml
openapi: 3.0.3
info:
  title: School Management API
  version: 1.0.0
  description: >
    API for school roles (super-admin, admin, teacher, student, parent),
    including attendance, scores, homework, leave requests, announcements,
    notifications, messages, and incident reports.
servers:
  - url: http://127.0.0.1:8000
security:
  - bearerAuth: []
paths:
...
  /api/scores:
    get:
      tags: [Scores]
      summary: List scores
      parameters:
        - in: query
          name: class_id
          schema: { type: integer }
        - in: query
          name: student_id
          schema: { type: integer }
        - in: query
          name: subject_id
          schema: { type: integer }
        - in: query
          name: month
          schema: { type: integer, minimum: 1, maximum: 12 }
        - in: query
          name: quarter
          schema: { type: integer, minimum: 1, maximum: 4 }
        - in: query
          name: period
          schema: { type: string }
      responses:
        '200':
          description: Scores pagination result
    post:
      tags: [Scores]
      summary: Create score (teacher/admin/super-admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [student_id, subject_id, class_id, exam_score, total_score]
              properties:
                student_id: { type: integer }
                subject_id: { type: integer }
                class_id: { type: integer }
                exam_score: { type: number }
                total_score: { type: number }
                month: { type: integer }
                quarter: { type: integer }
                period: { type: string }
                grade: { type: string }
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token